New IS160 Forensic Economics/requirements.txt
openpyxl>=3.1.2
pandas>=2.2.0
numpy>=1.26.0
requests>=2.31.0
beautifulsoup4>=4.12.2
lxml>=4.9.3
python-dateutil>=2.8.2
PyPDF2>=3.0.1
tabula-py>=2.9.0
flask>=3.0.0

